updated for dry run and for loop

+++ ansible/regtest/library/testcases/bgp/test_bgp_as_path.py
@@ -22,6 +22,8 @@ import shlex
 
 from collections import OrderedDict
 
+import time
+
 from ansible.module_utils.basic import AnsibleModule
 
 DOCUMENTATION = """
@@ -111,7 +113,7 @@ def execute_commands(module, cmd):
     """
     global HASH_DICT
 
-    if 'service' in cmd and 'restart' in cmd:
+    if 'service' in cmd and 'restart' in cmd or module.params['dry_run_mode']:
         out = None
     else:
         out = run_cli(module, cmd)
@@ -125,17 +127,56 @@ def execute_commands(module, cmd):
     return out
 
 
+def verify_bgp(module):
+    # Get all bgp routes
+
+    global RESULT_STATUS, HASH_DICT
+    failure_summary = ''
+    switch_name = module.params['switch_name']
+    config_file = module.params['config_file'].splitlines()
+    cmd = "vtysh -c 'sh ip bgp'"
+    bgp_out = execute_commands(module, cmd)
+
+    if bgp_out:
+        for line in config_file:
+            line = line.strip()
+            if line.startswith('ip prefix-list'):
+                ip = line.split().pop()
+                if ip not in bgp_out:
+                    RESULT_STATUS = False
+                    failure_summary += 'On switch {} '.format(switch_name)
+                    failure_summary += 'bgp route for network {} '.format(ip)
+                    failure_summary += 'is not present in the '
+                    failure_summary += 'output of command {}\n'.format(cmd)
+
+            if line.startswith('set as-path'):
+                as_path = line[-3::]
+                if as_path not in bgp_out:
+                    RESULT_STATUS = False
+                    failure_summary += 'On switch {} '.format(switch_name)
+                    failure_summary += 'set as-path is not present in the '
+                    failure_summary += 'output of command {}\n'.format(cmd)
+    else:
+        RESULT_STATUS = False
+        failure_summary += 'On switch {} '.format(switch_name)
+        failure_summary += 'bgp as path cannot be verified since '
+        failure_summary += 'output of command {} '.format(cmd)
+        failure_summary += 'is None'
+
+    alist = [True if RESULT_STATUS else False]
+    alist.append(failure_summary)
+    return alist
+
+
 def verify_bgp_as_path(module):
     """
     Method to verify bgp as path.
     :param module: The Ansible module to fetch input parameters.
     """
     global RESULT_STATUS, HASH_DICT
-    failure_summary = ''
     switch_name = module.params['switch_name']
     package_name = module.params['package_name']
     as_path_switch = module.params['as_path_switch']
-    config_file = module.params['config_file'].splitlines()
 
     # Get the current/running configurations
     execute_commands(module, "vtysh -c 'sh running-config'")
@@ -143,39 +184,19 @@ def verify_bgp_as_path(module):
     # Restart and check package status
     execute_commands(module, 'service {} restart'.format(package_name))
     execute_commands(module, 'service {} status'.format(package_name))
-
+    found_list = [False]
     if switch_name == as_path_switch:
-        # Get all bgp routes
-        cmd = "vtysh -c 'sh ip bgp'"
-        bgp_out = execute_commands(module, cmd)
-
-        if bgp_out:
-            for line in config_file:
-                line = line.strip()
-                if line.startswith('ip prefix-list'):
-                    ip = line.split().pop()
-                    if ip not in bgp_out:
-                        RESULT_STATUS = False
-                        failure_summary += 'On switch {} '.format(switch_name)
-                        failure_summary += 'bgp route for network {} '.format(ip)
-                        failure_summary += 'is not present in the '
-                        failure_summary += 'output of command {}\n'.format(cmd)
-
-                if line.startswith('set as-path'):
-                    as_path = line[-3::]
-                    if as_path not in bgp_out:
-                        RESULT_STATUS = False
-                        failure_summary += 'On switch {} '.format(switch_name)
-                        failure_summary += 'set as-path is not present in the '
-                        failure_summary += 'output of command {}\n'.format(cmd)
-        else:
-            RESULT_STATUS = False
-            failure_summary += 'On switch {} '.format(switch_name)
-            failure_summary += 'bgp as path cannot be verified since '
-            failure_summary += 'output of command {} '.format(cmd)
-            failure_summary += 'is None'
-
-    HASH_DICT['result.detail'] = failure_summary
+        retry = retries - 1
+        while(retry):
+            if not found_list[0]:
+                if verify_bgp(module)[0]:
+                    found_list[0] = True
+            if all(found_list):
+                break
+            else:
+                time.sleep(delay)
+                retry -= 1
+    HASH_DICT['result.detail'] = verify_bgp(module)[1]
 
     # Get the GOES status info
     execute_commands(module, 'goes status')
@@ -188,6 +209,9 @@ def main():
             switch_name=dict(required=False, type='str'),
             config_file=dict(required=False, type='str'),
             as_path_switch=dict(required=False, type='str'),
+            delay=dict(required=False, type='int', default=10),
+            retries=dict(required=False, type='int', default=6),
+            dry_run_mode=dict(required=False, type='bool', default=False),
             package_name=dict(required=False, type='str'),
             hash_name=dict(required=False, type='str'),
             log_dir_path=dict(required=False, type='str'),
@@ -195,30 +219,50 @@ def main():
     )
 
     global HASH_DICT, RESULT_STATUS
+    if module.params['dry_run_mode']:
+        cmds_list = []
+        package_name = module.params['package_name']
+        # Get the current/running configurations
+        execute_commands(module, "vtysh -c 'sh running-config'")
 
-    verify_bgp_as_path(module)
+        # Restart and check package status
+        execute_commands(module, 'service {} restart'.format(package_name))
+        execute_commands(module, 'service {} status'.format(package_name))
+        cmd = "vtysh -c 'sh ip bgp'"
+        execute_commands(module, cmd)
+        execute_commands(module, 'goes status')
+        for key, value in HASH_DICT.iteritems():
+            cmds_list.append(key)
 
-    # Calculate the entire test result
-    HASH_DICT['result.status'] = 'Passed' if RESULT_STATUS else 'Failed'
+        # Exit the module and return the required JSON.
+        module.exit_json(
+            cmds=cmds_list
+        )
+    else:
 
-    # Create a log file
-    log_file_path = module.params['log_dir_path']
-    log_file_path += '/{}.log'.format(module.params['hash_name'])
-    log_file = open(log_file_path, 'w')
-    for key, value in HASH_DICT.iteritems():
-        log_file.write(key)
-        log_file.write('\n')
-        log_file.write(str(value))
-        log_file.write('\n')
-        log_file.write('\n')
+        verify_bgp_as_path(module)
 
-    log_file.close()
+        # Calculate the entire test result
+        HASH_DICT['result.status'] = 'Passed' if RESULT_STATUS else 'Failed'
 
-    # Exit the module and return the required JSON.
-    module.exit_json(
-        hash_dict=HASH_DICT,
-        log_file_path=log_file_path
-    )
+        # Create a log file
+        log_file_path = module.params['log_dir_path']
+        log_file_path += '/{}.log'.format(module.params['hash_name'])
+        log_file = open(log_file_path, 'w')
+        for key, value in HASH_DICT.iteritems():
+            log_file.write(key)
+            log_file.write('\n')
+            log_file.write(str(value))
+            log_file.write('\n')
+            log_file.write('\n')
+
+        log_file.close()
+
+        # Exit the module and return the required JSON.
+        module.exit_json(
+            hash_dict=HASH_DICT,
+            log_file_path=log_file_path
+        )
 
 if __name__ == '__main__':
     main()
